requirements.txt
langchain-huggingface
transformers
pinecone-client 
streamlit
huggingface_hub
langchain-community
langchain
numpy
